+++ src/monitoring/prometheus_exporter.cpp
@@ -54,7 +54,7 @@ class SocketHandle {
   int fd_;
 };
 
-PrometheusExporter::PrometheusExporter(int port) : port_(port) {}
+PrometheusExporter::PrometheusExporter(uint16_t port) : port_(port) {}
 
 PrometheusExporter::~PrometheusExporter() {
   stop();
@@ -136,9 +136,7 @@ bool PrometheusExporter::isRunning() const {
   return running_.load();
 }
 
-int PrometheusExporter::getPort() const {
-  return port_;
-}
+uint16_t PrometheusExporter::getPort() const { return port_; }
 
 void PrometheusExporter::serverLoop() {
   while (running_.load()) {
